requirements.txt
pandas==2.2.2
psycopg2-binary==2.9.9
numpy==1.26.4
python-dateutil==2.9.0